Fixed first flutter test
